Struct enum_set::EnumSet

pub struct EnumSet<E> {
    // some fields omitted
}

A specialized set implementation to use enum types.

Methods

impl<E: CLike> EnumSet<E>

fn new() -> Self

Returns an empty EnumSet.

fn len(&self) -> usize

Returns the number of elements in the set.

fn is_empty(&self) -> bool

Checks if the set is empty.

fn clear(&mut self)

Removes all elements from the set.

fn is_disjoint(&self, other: &Self) -> bool

Returns true if the set has no elements in common with other.

This is equivalent to checking for an empty intersection.

fn is_superset(&self, other: &Self) -> bool

Returns true if the set is a superset of other.

fn is_subset(&self, other: &Self) -> bool

Returns true if the set is a subset of other.

fn union(&self, other: Self) -> Self

Returns the union of the set and other.

fn intersection(&self, other: Self) -> Self

Returns the intersection of the set and other.

fn difference(&self, other: Self) -> Self

Returns the difference between the set and other.

fn symmetric_difference(&self, other: Self) -> Self

Returns the symmetric difference between the set and other.

fn insert(&mut self, value: E) -> bool

Adds the given value to the set.

Returns true if the value was not already present in the set.

fn remove(&mut self, value: &E) -> bool

Removes a value from the set.

Returns true if the value was present in the set.

fn contains(&self, value: &E) -> bool

Returns true if the set contains the given value.

fn iter(&self) -> Iter<E>

Returns an iterator over the set's elements.