Struct enum_set::EnumSet

pub struct EnumSet<E> { /* fields omitted */ }

A specialized set implementation to use enum types.

Methods

impl<E: CLike> EnumSet<E>

pub fn new() -> Self

Returns an empty EnumSet.

pub fn len(&self) -> usize

Returns the number of elements in the set.

pub fn is_empty(&self) -> bool

Checks if the set is empty.

pub fn clear(&mut self)

Removes all elements from the set.

pub fn is_disjoint(&self, other: &Self) -> bool

Returns true if the set has no elements in common with other.

This is equivalent to checking for an empty intersection.

pub fn is_superset(&self, other: &Self) -> bool

Returns true if the set is a superset of other.

pub fn is_subset(&self, other: &Self) -> bool

Returns true if the set is a subset of other.

pub fn union(&self, other: Self) -> Self

Returns the union of the set and other.

pub fn intersection(&self, other: Self) -> Self

Returns the intersection of the set and other.

pub fn difference(&self, other: Self) -> Self

Returns the difference between the set and other.

pub fn symmetric_difference(&self, other: Self) -> Self

Returns the symmetric difference between the set and other.

pub fn insert(&mut self, value: E) -> bool

Adds the given value to the set.

Returns true if the value was not already present in the set.

pub fn remove(&mut self, value: &E) -> bool

Removes a value from the set.

Returns true if the value was present in the set.

pub fn contains(&self, value: &E) -> bool

Returns true if the set contains the given value.

Important traits for Iter<E>
impl<E: CLike> Iterator for Iter<E> type Item = E;

pub fn iter(&self) -> Iter<E>

Returns an iterator over the set's elements.